@@ -1,8 +1,8 @@
 package com.github.windsekirun.bindadapters.sample
 
-import android.databinding.DataBindingUtil
+import androidx.databinding.DataBindingUtil
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.github.windsekirun.bindadapters.sample.databinding.MainActivityBinding
 
 class MainActivity : AppCompatActivity() {
