@@ -1,8 +1,8 @@
 package com.github.windsekirun.bindadapters.widget
 
-import android.databinding.BindingMethod
-import android.databinding.BindingMethods
-import android.databinding.InverseBindingAdapter
+import androidx.databinding.BindingMethod
+import androidx.databinding.BindingMethods
+import androidx.databinding.InverseBindingAdapter
 import android.view.View
 import android.widget.EditText
 
