@@ -1,12 +1,12 @@
 package com.github.windsekirun.bindadapters.sample
 
-import android.arch.lifecycle.Lifecycle
-import android.arch.lifecycle.LifecycleObserver
-import android.arch.lifecycle.OnLifecycleEvent
-import android.databinding.ObservableArrayList
-import android.databinding.ObservableField
-import android.databinding.ObservableInt
-import android.databinding.ObservableList
+import androidx.lifecycle.Lifecycle
+import androidx.lifecycle.LifecycleObserver
+import androidx.lifecycle.OnLifecycleEvent
+import androidx.databinding.ObservableArrayList
+import androidx.databinding.ObservableField
+import androidx.databinding.ObservableInt
+import androidx.databinding.ObservableList
 import android.net.Uri
 import android.view.View
 import com.github.windsekirun.bindadapters.observable.ObservableString
